Overall refactoring for passing key presses from kernel to apps

+++ include/p_kernel.h
@@ -13,7 +13,7 @@ extern uint16_t krn_event_count(void);
 extern void far * krn_heap_alloc(uint16_t size);
 extern void krn_heap_init(void);
 /* kernel/keyboard.c */
-extern uint8_t krn_keyboard_getc(void);
+extern uint16_t krn_keyboard_getc(void);
 extern void krn_keyboard_handle_intr(void);
 extern void krn_keyboard_init(void);
 extern void krn_keyboard_deinit(void);
